+++ src/main/java/gift/dto/ProductRequestDto.java
@@ -1,4 +1,4 @@
-package gift;
+package gift.dto;
 
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
